Creating main file.
